+++ src/main/kotlin/hd/junction/visit/dto/request/VisitRequestDto.kt
@@ -1,0 +1,23 @@
+package hd.junction.visit.dto.request
+
+import hd.junction.patient.dto.request.PatientRequestDto
+import hd.junction.patient.dto.request.VisitPatientRequestDto
+import java.time.LocalDate
+
+data class VisitRequestDto(
+    val reservationDate: LocalDate,
+
+    val visitStateCode: String,
+
+    val hospitalId: Long,
+
+    val patientId: Long? = null,
+
+    val patientRequest: VisitPatientRequestDto? = null,
+) {
+    init {
+        require(patientId != null || patientRequest != null) {
+            "환자 ID 또는 환자 정보 중 하나는 반드시 제공되어야 합니다"
+        }
+    }
+}